requirements.txt
PyAudio == 0.2.11
Pillow == 7.0.0
SpeechRecognition == 3.8.1
beepy == 1.0.7
pyspellchecker == 0.6.2
pyttsx3 == 2.90